pyproject.toml
[project]
name = "dds-ai-application-challenge"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "groq>=1.2.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
